Changed stats floats to ints

+++ Assets/_Assets/Scripts/Player/WeaponDetect.cs
@@ -47,7 +47,7 @@ public class WeaponDetect : MonoBehaviour
             currentDelay = Time.time + 0.5f; // Enemy invincible for 0.5secs
 
             enemy.TakeDamage(damage);
-            gameManager.dmgDealt += damage;
+            gameManager.dmgDealt += (int)damage;
         }
     }
 }
